refactor: move components appearance setup to web app
Turns out it was a bad idea to move the entire appearance module to the components' directory.
Since the electron setup is a bit different and way simpler than the browser setup.
So moving all the stuff back to the web app where it is used only.

Sure there are a few duplications in each setup, but okay for now.
Maybe we can optimize later, let's see.